+++ provider/cmd/crd2pulumi/gen/utilities.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package schema
+package gen
 
 import (
 	"bytes"
@@ -21,6 +21,7 @@ import (
 	"io/ioutil"
 	"strings"
 
+	"github.com/pulumi/pulumi/sdk/go/common/util/contract"
 	unstruct "k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/util/yaml"
 )
@@ -46,7 +47,7 @@ func NestedMapSlice(obj map[string]interface{}, fields ...string) ([]map[string]
 	}
 	m, ok := val.([]interface{})
 	if !ok {
-		return nil, true, fmt.Errorf("%v accessor error: %v is of the type %T, expected []interface{}", jsonPath(fields), val, val)
+		return nil, false, fmt.Errorf("%v accessor error: %v is of the type %T, expected []interface{}", jsonPath(fields), val, val)
 	}
 	mapSlice := make([]map[string]interface{}, 0, len(m))
 	for _, v := range m {
@@ -59,8 +60,18 @@ func NestedMapSlice(obj map[string]interface{}, fields ...string) ([]map[string]
 	return mapSlice, true, nil
 }
 
+func IsValidLanguage(language string) bool {
+	return language == NodeJS || language == Go || language == Python || language == DotNet
+}
+
 func jsonPath(fields []string) string {
 	return "." + strings.Join(fields, ".")
+}
+
+func rawMessage(v interface{}) json.RawMessage {
+	bytes, err := json.Marshal(v)
+	contract.Assert(err == nil)
+	return bytes
 }
 
 // GenericizeStringSlice converts a []string to []interface{}.